x11-fonts/noto-serif-tc: Add noto-serif-tc 2.001
Noto Sans CJK and Noto Serif CJK comprehensively cover Simplified Chinese,
Traditional Chinese, Japanese, and Korean in a unified font family. This
includes the full coverage of CJK Ideographs with variation support for 4
regions, Kangxi radicals, Japanese Kana, Korean Hangul, and other CJK symbols
and letters in the Basic Multilingual Plane of Unicode. It also provides limited
coverage of CJK Ideographs in Plane 2 of Unicode as necessary to support
standards from China and Japan.

This port is the Serif set of Traditional Chinese fonts which supports BIG5.
Traditional Chinese glyphs comply with the glyph standard of the Taiwan Ministry
of Education.