misc/qt6-examples: new port
A collection of code samples and tutorials are provided with Qt to help
new users get started with Qt development. These documents cover a range
of topics, from basic use of widgets to step-by-step tutorials that show
how an application is put together.

The update to 6.4.2 was provided by Florian Walpen <dev@submerge.ch>.

PR:		269028