net/quagga: use MORE=-E implicitly within vtysh
quagga's vtysh utility defaults to "more" as PAGER that is less(1)
under FreeBSD and its current behaviour is not suitable for vtysh.

Generally, setting MORE=-E in user environment makes it behave better
but due to collateral damage in bash version 5+
this setting may be ignored.

Let's make it work out of the box while still allowing users to change or
restore pager behaviour with custom MORE settings.