irc/kvirc: update to latest commit
There's a handful of bugfixes and compatibility improvements
in the not-otherwise-released 5.0 (main) branch of kvirc.
Update anyway, since at least one is a crash fix.

Chris proposed an update to -85, but I looked at the rest
of the newer commits and concluded they would be useful too.
So some of this is Chris's work, some is my own.

PR:		267956