devel/llvm*: Fix CONFLICT on cmark
If the user has the textproc/cmark package installed, the py*-CommonMark
package will fail to install.  This in turn leads to build failures down
the line when sphinx tries to load commonmark.  The result is confusing
and potentially wastes a lot of build time so just conflict upfront.
The previous code attempted to do this, but I'd listed the conflict
incorrectly in 2019. :(

PR:		239636
Reported by:	Stephen Roznowski <sjrbulk@verizon.net>