sysutils/tracker: fix build with clang 15
During an exp-run for llvm 15 (see bug 265425), it turned out that
sysutils/tracker failed to build with clang 15:

  src/tracker-store/tracker-store.p/tracker-resources.c:1685:14: error: incompatible pointer to integer conversion assigning to 'gint' (aka 'int') from 'gconstpointer' (aka 'const void *') [-Wint-conversion]
                  subject_id = _tmp1_;
                             ^ ~~~~~~

This is because in this vala-generated C code, the glib hash table
iterator code always returns pointers, even for integer keys. However,
vala is inflexible in the C expressions it generates to store the
outcome. Suppress -Wint-conversion to get rid of the compile error.

PR:		268787
Approved by:	tcberner (maintainer)
MFH:		2023Q1