graphics/gimp-app: fix build with clang 15
During an exp-run for llvm 15 (see bug 265425), it turned out that
graphics/gimp-app failed to build with clang 15:

  file-dicom.c:605:26: error: incompatible pointer to integer conversion returning 'void *' from a function with result type 'gint32' (aka 'int') [-Wint-conversion]
                    return NULL;
                           ^~~~

This is because the return statement is a function that should return an
integer, not a pointer. Replace it with the error value -1, as used
elsewhere in the function.

PR:		268790
Approved by:	fluffy (maintainer)
MFH:		2023Q1