x11/mate-panel: fix build with clang 15
During an exp-run for llvm 15 (see bug 265425), it turned out that
x11/mate-panel failed to build with clang 15:

  clock-location.c:454:22: error: incompatible pointer to integer conversion assigning to 'glong' (aka 'long') from 'char *(int, int)' [-Wint-conversion]
          sys_timezone = timezone;
                       ^ ~~~~~~~~
  clock-location.c:462:24: error: incompatible pointer to integer conversion assigning to 'glong' (aka 'long') from 'char *(int, int)' [-Wint-conversion]
          local_timezone = timezone;
                         ^ ~~~~~~~~

This is because 'timezone' as a global external variable does not exist
on FreeBSD: it is a glibc-ism. Use struct tm's tm_gmtoff field instead,
which has its sign reversed from the glibc global. There is also no need
to manually account for DST, as tm_gmtoff includes that.

PR:		268795
Approved by:	portmgr (tcberner)
MFH:		2023Q1