dns/dnsmasq: fix rare cache corruption problem
Simon Kelley sent an advisory that in rare circumstances, the cache can
become corrupted and the DNS subsystem then became disfunctional.
This is reported as regression in 2.88.
Chances seem higher this happens with DNSSEC enabled, but seems not limited
to it.  For details, please see the patch contained in this commit, or
https://lists.thekelleys.org.uk/pipermail/dnsmasq-discuss/2023q1/016821.html

The symptom of this happening is apparently a cache internal error.

2.89 with this fix and a few others is slated for release in a week.
Let's fix the patch already and MFH to 2023Q1 so we keep our liberties
to decide whether we need to move quarterly to 2.89 or rather stick
with 2.88_1.

originally
Reported by:	Timo van Roermund (to Simon Kelley in private)
Reported by:	Simon Kelley (upstream maintainer, through mailing list)
Obtained from:	Simon Kelley (upstream maintainer, Git repository)
MFH:		2023Q1