net/pload: Fix command-line args and unit display.
This patch updates the port to fix the fact that the code calls
XtAppInitialize() which eats the command-line args and so causes ps(1)
to not display them. The command-line args are now saved before and restored
after the call to XtAppInitialize().

Also, correct the display of the units of total byte counts to show as "B",
"kB", "GB" etc, rather than the imprecise/incorrect "b", "k", "G" etc.

PR:		267710
Approved by:	J.R. Oldroyd (maintainer)