archivers/rpm4: re-enable the build of the bundled elfdeps utility
Commit 5ff99a4c9df9 disabled libelf detection to work around the build
failure, but rpmbuild expects to be able to find the elfdeps utility and
misbehaves when it's not present.  Since the build failure is easy to
address (different names for an ELF machine type constant for the rather
obsolescent DEC Alpha), let's just do that.

As a part of this, elfutils must be listed as a runtime dependency.
Note that we cannot list it in LIB_DEPENDS since the ports framework
ends up finding the base system libelf.so and decides that the
dependency is satisifed without installing elfutils.  But the base
system libelf is missing a symbol (gelf_getvernaux) required by elfdeps.
Work around this by listing elfutils as both build and runtime
dependencies.

Approved by:	rodrigo
Reviewed by:	rodrigo
Sponsored by:	Klara, Inc.
Differential Revision:	https://reviews.freebsd.org/D38092