net/kube-apiserver: Add new port
The Kubernetes API server validates and configures data for the api
objects which include pods, services, replicationcontrollers,
and others. The API Server services REST operations and provides
the frontend to the cluster's shared state through which all other
components interact.

Note: while testing this as a replacement of a Linux based on my lab,
I only saw Webhooks not working due to Kubernetes network mesh not implemented.
I added a notice about this for people who want to test it.

PR:		267308