games/chromono: add port: Circular color puzzle
All those half-colored spheres could really need some help. And
with that, we obviously mean your help, otherwise this would probably
not be a game, but a movie or something. Grab those fully-colored
spheres and share the color (and the correct one at that) to bring
the world order back to normal. Sounds easy? It is. Up until the
third level or so. Then it gets tricky. But you'll figure it out,
I'm sure.