graphics/darktable: Remove several options previously enabled by default
They all contribute to the full experience of Darktable and I never take
the time to test the port without these options. Let's simplify the port
then.

The only default option left is `PRINT`. Many users are probably not
printing their photos, so leave the possibility to turn this off.