lang/ruby31: Add upstream patch to fix bug of mkmf library
There is bug in mkmf library of Ruby 3.1 that it fails to detect
pkg-config command if it isn't installed in environment that Ruby is
built, and it causes build error of some ports when default version of
Ruby is 3.1. So add upstream patch to fix it.

Reference:	https://bugs.ruby-lang.org/issues/19189
PR:		269197
Approved by:	maintainer timeout
MFH:		2023Q1