devel/py-x-wr-timezone: Add py-x-wr-timezone 0.0.5
Some calendar providers introduce the non-standard X-WR-TIMEZONE parameter to
ICS calendar files. Strict interpretations according to RFC 5545 ignore the
X-WR-TIMEZONE parameter. This causes the times of the events to differ from
those which make use of X-WR-TIMEZONE.

This module aims to bridge the gap by converting calendars using X-WR-TIMEZONE
to a strict RFC 5545 calendars. So, let's put our heads together and solve this
problem for everyone!