devel/arcanist-lib: Suppress certain runtime exceptions
Default configured PHP 8.1+ issues deprecation errors when NULL
strings are passed into various library functions.

In the past, we tried to address those one by one, which feels
a bit like playing Whac-A-Mole.

This patch adds custom error handling for this kind of deprecation
error, so that it is ignored. This should fix this entire class of
errors at least until PHP 9.0 is released, at which point a different
workaround (or real solution is needed). Therefore previous patches
were removed.