science/py-cfgrib: Add py-cfgrib 0.9.10.3
cfgrib provides a Python interface to map GRIB files to the Unidata's Common
Data Model v4 following the CF Conventions. The high level API is designed to
support a GRIB engine for xarray and it is inspired by netCDF4-python and
h5netcdf. Low level access and decoding is performed via the ECMWF ecCodes
library and the eccodes python package.