devel/build2: Toolchain for building and packaging C++ code
build2 is an open source (MIT), cross-platform build toolchain that aims
to approximate Rust Cargo's convenience for developing and packaging
C/C++ projects while providing more depth and flexibility, especially in
the build system.

libodb and libodb-sqlite are currently bundled in this port.  These are
planned to be unbundled in a future commit.

The port's author is known to the maintainer but wishes not to be named.

Obtained from:	anonymous author
Approved by:	flo (mentor)
Differential Revision: https://reviews.freebsd.org/D38629