security/openssl-agent: New port: OpenSSL key agent and client utils
OpenSSL key agent and client utilities.

The aim of these utilities is to provide an openssl-rsautl(1) drop-in
replacement for performing cryptographic operations using a private key
that is unlocked for the session, similar to OpenSSH's ssh-agent(1).

The port's author is known to the maintainer but wishes not to be named.

WWW: https://git.build2.org/cgit/openssl-agent/tree/README

Obtained from:	anonymous author
Approved by:	flo (mentor)
Differential Revision: https://reviews.freebsd.org/D38630