lang/mlkit: Standard ML Compiler
The MLKit is a compiler toolkit for the Standard ML language, including
The MLKit with Regions, which features a native backend for
the x64 architecture, based on region inference, and SMLtoJs,
which features a JavaScript backend targeting web browsers.
The two compilers share the same frontend and compilation management scheme.

The MLKit covers all of Standard ML, as defined in the 1997 edition of
The Definition of Standard ML and supports most of
the Standard ML Basis Library.

WWW: https://elsman.com/mlkit/

PR:		269566