sysutils/e2fsprogs-core: clean up a bit, drop coreutils depends
Patch a test script to work around BSD dd's missing oconv=append,
and do away with coreutils as a build requisite.

Move some REINPLACE_CMD features to individual patches that I
intend to send upstream for inclusion, and remove others that
are now no longer required.

This should only affect the build and self-test phases,
and not change package content, so not bumping PORTREVISION.