games/gzdoom: cleanup dependencies
Most of the GNOME and the GL dependencies are transitive from GTK, not
directly used by gzdoom -- axe them.

While we're here, switch to GTK 3.  gzdoom already supports it out of
the box and GTK 2 is EOL, so the timing is right.  DYN_GTK has been
switched off because I like to see that the dependency's actually in
use at build time.