devel/tcl-trf: fix fatal compiler warning
Starting with version 15 clang includes the new -Wdeprecated-non-prototype
in -Wall. This warns, when a K&R-style function-definitions follow
ANSI function-declarations. And TCL code has LOTS of K&R definitions...

The warning is poorly worded -- and, in my opinion, outright bogus,
but the ports need to keep building, so I'm using a trick borrowed
from src/share/mk/sys.mk to conditionally disable it.