textproc/wordnet: fix newly-flagged compiler warnings
clang15 used by current is stricter -- probably, incorrectly so.
For example, -Wstrict-prototypes will now flag the following:

	int fun(void);

	...

	int fun()
	{
	}

The definition must have an explicit "void"... There was also a
variable, which was set -- and even incremented -- but the of its
value was commented out in the upstream code decades ago. The
compiler now flags it -- correctly -- as set, but not used.