emulators/qemu: Fix missing LIB_DEPENDS
Add net/libslirp into LIB_DEPENDS, now required for making user
networking functional.

In qemu 7.2 release the slirp code has been moved out of the source
package and is enabled/disabled at the build time based on the
availability of the libslirp shared library. As such, it needs to be
added into LIB_DEPENDS to make this feature available in the quemu
package.

See
https://wiki.qemu.org/ChangeLog/7.2#Removal_of_the_.22slirp.22_submodule_.28affects_.22-netdev_user.22.29
for a reference.

PR:		269475
Reported by:	sobomax