base/*: Remove the base ports to install an external toolchain in /usr.
This was originally motivated for providing a native toolchain for
platforms not supported by the in-tree LLVM toolchain.  Now, however,
all of our current platforms are well supported by the in-tree LLVM
toolchain.  In addition, these ports required additional maintenance
beyond the existing devel/binutils and devel/freebsd-gccX ports.  A
more sustainable path if we need this in the future is probably to
have simple wrappers for /usr/bin/cc and other binaries in /usr that
redirect to binaries provided by the existing devel/ ports.

Reviewed by:	bapt
Differential Revision:	https://reviews.freebsd.org/D38561