devel/libphonenumber: update to 8.13.6, latest upstream
Release notes are at
	https://github.com/google/libphonenumber/releases/tag/v8.13.6

This change to ports is a little larger than usual: the newer releases
want to regenerate metadata (which in turn is needed to run the tests)
but the (re)generation code is Java, so building that would pull in
an entire JRE. I claim that we have enough big things in the build
chain already, so that means disabling building-the-tests and
not regenerating the metadata. It's a bit unclear whether this
is sustainable: the metadata looks like it is transitioning to CSV,
which I would expect to be *simpler* in use, rather than requiring
even more heavyweight tools.