www/caddy: Update to 2.6.4
Changes:
This release contains a hotfix for a regression in v2.6.3 related to
proxying chunked requests. We recommend that all users who do so upgrade
to v2.6.4.

Note that, in an effort to make error-prone configs less likely, we have
deprecated the reverse proxy options:

    buffer_requests
    buffer_responses
    max_buffer_size

and have introduced 2 new ones which take a size argument to enable
buffering:

    request_buffers <size>
    response_buffers <size>

The deprecated options will be removed in a later version of Caddy, so
please start using the new parameters instead.