deskutils/calibre: Update to 6.12.0
With this release upstream has decided not to include some prebuilt
and bundled parts. So the port now needs to download some extra
files, to be able to compile.

Also while building calibre now needs to launch parts of itself
which creates a temporary directory in LOCALBASE, if it is writable,
so clean that up after the build.

- Use DIST_SUBDIR, since the port is downloading multiple files
  now.
- Add new dependencies for things not embedded anymore
- Remove unneeded EXTRACT_BEFORE_ARGS
- Remove references to python2
- Bundle predigested user-agent-data.json file, adding a utility
  target to easily regenerate it
- Move from dead upstream textproc/cchardet to drop-in-ready
  textproc/py-charset-normalizer [1]

Submitted by:	fluffy [1]