ftp/ftpmirror: Deprecate and reset maintainership
This project is old and has no development for a long time.  FTP is not
that relevant nowadays and recently it stopped building.

I have no interest on it and will not put effort to get it fixed, so I
dropped maintainership and marked it to be removed in 3 months.