multimedia/transcode: mark as broken and deprecated
This software has been abandoned by the author and is not longer
available from an official repository. A verbatim copy exists on
GitHub, and the port is modified to be fetched from that repository.

This port is broken since it depends on audio resample functions that
have been deprecated in 2013 and removed from the ffmpeg sources in
2017, and therefore does not build.

The port is marked deprecated and to be removed after 2023-03-31.
If somebody wants to fix the port I'll accept patches (or transder
maintainership, since I do not use this software anymore).

Reported by:	pkg-fallout