net/rabbitmq: rc.d improvements
This commit creates a separate directory at /var/run/rabbitmq to
host the RabbitMQ pid file. Previously, the pid file was not being
created properly (it was empty) and this was causing issues with
the process of waiting for pids (as the pid file existed, but had
no value and was considered "garbage"). With this change the pid
file is created with the correct value, enabling the process of
waiting for pids to complete successfully; now the startup script
returns exit code 0.

Additionally, this commit exports the variables RABBITMQ_HOME,
RABBITMQ_LOG_BASE, and RABBITMQ_PID_FILE so that the RabbitMQ process
can properly locate the pid file.

These changes resolve PR 264062.

PR:		264062
Reported by:	FiLiS <freebsdbugs@filis.org>