math/readstat: fix build on 14-CURRENT
In 14-CURRENT, the system compiler is more strict; it looks like

  In file included from src/readstat_bits.c:9:
    src/readstat_bits.h:10:29: error: a function declaration without
    a prototype is deprecated in all versions of C
    [-Werror,-Wstrict-prototypes]
  int machine_is_little_endian();

Apply the stricter flags in all builds, then fix the source so
that the prototypes are explicitly (void) -- this is a subtlely
of C semantics where (void) means "no arguments" and () means
"arguments unspecified" that rarely matters. Except now.