Mark ports that depend on transcode as deprecated
The multimedia/transcode port is broken due to its dependency on a
long obsolete version of ffmpeg (specifically the audio resample
functions). It is unmaintained and the original author has deleted
the sources and documentation.

Since I do not have any use for these ports and it would requiee
a lot of effort to port transcode to the new resample API, I'm not
going to try to modernize this software.

If anybody wants to take over maintainership of the transcode port
or dependent ports, feel free to do so. Else I'm going to delete
them after 2023-03-31.