x11/wscreensaver: add new port
This is an experimental fork of xscreensaver which makes it possible to run
some of the hacks as animated wallpapers on Wayland compositors that support
wlr-layer-shell.

With swaylock-plugin, it is possible to use these as backgrounds for a
lock screen on some Wayland compositors as well.

Note: this is a very rough work in progress, so don't be surprised if half of
the compiled hacks fail to run, or render incorrectly.

https://git.sr.ht/~mstoeckl/wscreensaver