databases/py-aiosqlite: Fix build
The previous commit was done by mistake (some work in progress leftover
and forgotten that got committed after a poudriere build that not build
this port succeeded).

Fix the port to make it buildable again.  No need to bump PORTREVISION
since the port was completely unbuildable.

Pointy hat to:	me