devel/py-tendo: Add new port
The main feature of interest is a single process instance feature which
can be used to ensure only one proces runs at a time, given multiple
users. Implemented via fnctl "under the hood" to create/check/remove
a lock file. While on FreeBSD we have lockf(1), this is not portable.