net/wireshark:  Update to 4.0.4 and refactor the port logic
See https://www.wireshark.org/docs/relnotes/wireshark-4.0.4.html for a
list of changes in wireshark.

This port features the following modifications:

- Use of helpers and consistent usage
- Refresh dependency list
- Use (lib)speexdsp from tree
- Add brotli as dependency, used for compressing data over HTTP2
- Flavourize port and create two variants, one including UI and one without (-nox11)
- Disable detection of git for Wireshark's build
- Rework and consolidate options, try to be as consistent as possible with the rest of the tree
- Rename DECRYPT option to GNUTLS
- Move various *dump tools to one option as they share the same dependency (libssh)
- Make LTO default option on aarch64 and amd64
- RTP audio codecs are only used with GUI so make all use the same toggle

Thanks to diizzy for his refactoring work.  One thing to note for uses
of the -lite ports is that they are going away.  Flavors should be used
to build wireshark with the GUI (default) or without the GUI (nox11).

PR:		269103
Submitted by:	diizzy [port refactoring]