security/vuxml: Autofill CVE information
The `newentry` target accepts an optional parameter CVE_ID.
When provided, the newentry.sh script tries to retrieve information from the
NVD and MITRE databases and fill the template accordingly.

The script needs `textproc/jq` and warns the user and exists if it is not found.

How to use it:

make newentry CVE_ID=CVE-2022-39282

Note that this is just a helper. *YOU HUMAN* have to check that the information
is correct.

Reviewed by: tcberner, jlduran_gmail.com, mat
Differential Revision: https://reviews.freebsd.org/D38894