devel/msgpack-{c,cxx}: Update to 6.0.0
Remember that time two months ago when upstream suddenly changed their
library/cmake/header names and didn't tell anybody, or set a deprecation
period, or make any sort of warning or announcement of any kind? And
then I had to spent a ton of time fixing the breakages?

Well, apparently the msgpack team thought it was such a good idea that
they did it again. With no warning. And no deprecation period. And no
announcements.

I'm over it. I'm not chasing their renames across the tree to satisfy
upstream's two-second attention span. I'm installing links from all the
old names and we can remove them when all the consumers switch to the
new locations, which is in approximately never.