net-im/libquotient: try to fix build on 12.*, 13.0
This just avoids trying to build tests -- which fail with the error

/usr/include/c++/v1/__functional/perfect_forward.h:77:9: error: no
matching constructor for initialization of 'tuple<void (TestManager::*)(),
TestManager *>'

on 12.* and 13.0 and possibly 13.1 but not my local poudriere host,
which reports that it is 13-STABLE (from a month back or more). The
underlying problem seems to be tied to the LLVM / Clang STL that is
available. Patch from Eric, I tightened the bounds a bit.

Reported by:	Eric Camachat