www/gitlab-ce: fix a problem in the asset compile step
Make sure that files having an executable bit keep this executable bit.
This fixes a problem of the last step in the asset compile step fails
due to missing executable permission on a script.
It does not have any impact on the installation itself as the last step is just failing
which is just a check which is done after all assets are compiled.

Thanks a lot to @tz and @moin pointing to it and helping to understand the
source for it.