x11-wm/swayfx: add new port
Sway is an incredible window manager, and certainly one of the most
well established wayland window managers. However, it is restricted to
only include the functionality that existed in i3. This fork ditches
the simple wlr_renderer, and replaces it with our fx_renderer, capable
of rendering with fancy GLES2 effects. This, along with a couple of
minor changes, expands sway's featureset to include the following:

+ Anti-aliased rounded corners, borders, and titlebars
+ Shadows
+ Dim unfocused windows
+ Per application saturation control: Allows the user to set the
  saturation (Digital Vibrance) for specific applications. Great for
  some FPS games!
+ Scratchpad treated as minimize: Allows docks, or panels with a
  taskbar, to correctly interpret minimize / unminimize requests

https://github.com/WillPower3309/swayfx/