www/caddy{,-custom}: Fix startup with JSON configuration
Caddy supports a wide number of configuration file formats, including
its built-in "Caddyfile" format, and the "native" format is JSON.

When anything other than the native JSON format is used, caddy needs to
be passed the --adapter [formatname] flag (though it automatically
handles the case where the config filename is Caddyfile). However, as
JSON is the native format, there is no "json" adapter.

The added patch drops the --adapter flag when caddy_adapter is "json".

PR:		270120