textproc/sphinxsearch: return to the pool
Upstream switched to a commercial license years ago, and so we cannot
include more recent versions in the ports.  Despite indications on the
website that they would switch back to GPL, this hasn't happened in
more than 5 years, so I guess it is never going to happen.  From an
opensource point of view, this software is essentially abandonware.

However, given there is an actively developed upstream project, and
that project does provide FreeBSD binaries, it doesn't feel right to
deprecate the software entirely.