editors/openoffice-*: test for required jar file
Test for a required jar file that is optionally installed by
textproc/lucene and fail at the configure step with a helpful message
if we don't find it.  If the jar file is missing, javac fails during
the build of l10ntools module with an unhelpful error message that does
not pinpoint the cause of the problem.
PR:		270158
Reported by:	Ken Gunderson <kgunders@teamcool.net>